add gat service

+++ core/mappers/drug_interaction_mapper.py
@@ -1,3 +1,4 @@
+import numpy as np
 from tqdm import tqdm
 
 from core.repository_models.training_drug_interaction_dto import TrainingDrugInteractionDTO
@@ -10,7 +11,7 @@ def map_training_drug_interactions(query_results) -> list[TrainingDrugInteractio
         (drug_1_id, drugbank_id_1, drug_2_id, drugbank_id_2, interaction_type) = result
         interaction_entity = TrainingDrugInteractionDTO(drug_1=drug_1_id,
                                                         drug_2=drug_2_id,
-                                                        interaction_type=interaction_type)
+                                                        interaction_type=np.int16(interaction_type))
         interactions.append(interaction_entity)
 
     return interactions
